auth_service/requirements.txt
flask
psycopg2-binary
python-dotenv
bcrypt
pyjwt
flask-cors